fix: CourseSectionDao 컴파일 오류 수정
- item() 메서드 호출 시 int 타입 명시적 캐스팅 추가
- "reference to item is ambiguous" 오류 해결

+++ src/dao/CourseSectionDao.java
@@ -33,7 +33,7 @@ public class CourseSectionDao extends DataObject {
 
 		DataSet cllist = courseLesson.find("course_id = " + newCourseId + " AND status != -1");
 		while(cllist.next()) {
-			courseLesson.item("section_id", sectionMap.containsKey(cllist.i("section_id")) ? sectionMap.get(cllist.i("section_id")) : 0);
+			courseLesson.item("section_id", (int)(sectionMap.containsKey(cllist.i("section_id")) ? sectionMap.get(cllist.i("section_id")) : 0));
 			if(!courseLesson.update("course_id = " + newCourseId + " AND lesson_id = " + cllist.i("lesson_id"))) return false;
 		}
